Relocate batch 1 of per-estimator configs into helper packages
Move 10 live per-estimator config classes out of config_models.py into
their estimator's utils/<name>_helpers/config.py, re-exported via the
existing lazy __getattr__ shim so imports such as
`from mlsynth.config_models import FMAConfig` keep working unchanged:

  FMA, PDA, CLUSTERSC, DSC, MUSC, MASC, SpSyDiD, MCNNM, MSQRT, SSC

Each relocated module carries only the imports its body uses. No behavior
change. config_models.py: 4218 -> 3468 lines.
