+++ mlsynth/config_models.py
@@ -339,101 +339,6 @@ class BaseEstimatorConfig(BaseModel):
 
 # Placeholder for other estimator configs - to be added sequentially
 
-class FMAConfig(BaseEstimatorConfig):
-    """Configuration for the Factor Model Approach (FMA) estimator.
-
-    Implements Li & Sonnier (2023), *"Statistical Inference for the
-    Factor Model Approach to Estimate Causal Effects in
-    Quasi-Experimental Settings"*, JMR 60(3):449-472. FMA estimates
-    the ATT for a single treated unit by extracting principal-
-    component factors from the control panel, projecting the treated
-    unit's pre-period outcomes onto those factors, and using the
-    fitted loading to predict the untreated potential outcome in the
-    post-period.
-
-    Parameters
-    ----------
-    stationarity : {"stationary", "nonstationary"}
-        Selects the factor-selection criterion: ``"stationary"`` uses
-        the paper's modified Bai-Ng (MBN) criterion (Web Appendix
-        D.1); ``"nonstationary"`` uses Bai (2004) IPC1 with a
-        log-log adjustment for non-stationary factors. Default:
-        ``"nonstationary"`` (the paper's recommendation for general
-        applied settings).
-    preprocessing : {"demean", "standardize"}
-        Preprocessing applied to the control panel before PCA.
-    n_factors : int or None
-        Override the data-driven factor count. ``None`` triggers the
-        criterion in ``stationarity``.
-    max_factors : int
-        Upper bound passed to the factor-selection routine.
-    alpha : float
-        Two-sided significance level for CIs.
-    inference_methods : list of {"asymptotic", "bootstrap", "placebo"}
-        Inference procedures to run. Defaults to ``["asymptotic"]``,
-        which gives the paper's Theorem 3.1 normal CI for the ATT.
-        Add ``"bootstrap"`` to get per-period ATT_t CIs via the Web
-        Appendix F residual bootstrap, and ``"placebo"`` to get the
-        Web Appendix G control-as-pseudo-treated band.
-    n_bootstrap : int
-        Number of bootstrap replicates (Web Appendix F). Ignored when
-        ``"bootstrap"`` is not in ``inference_methods``.
-    bootstrap_seed : int
-        Seed for the bootstrap RNG.
-    """
-
-    stationarity: Literal["stationary", "nonstationary"] = Field(
-        default="nonstationary",
-        description="Stationarity assumption for factor selection.",
-    )
-    preprocessing: Literal["demean", "standardize"] = Field(
-        default="demean",
-        description="Preprocessing applied to the control panel before PCA.",
-    )
-    n_factors: Optional[int] = Field(
-        default=None, ge=1,
-        description="Optional override of the data-driven factor count.",
-    )
-    max_factors: int = Field(
-        default=10, ge=1,
-        description="Upper bound on the factor-selection routine.",
-    )
-    alpha: float = Field(
-        default=0.05, gt=0.0, lt=1.0,
-        description="Two-sided significance level for CIs.",
-    )
-    inference_methods: List[str] = Field(
-        default_factory=lambda: ["asymptotic"],
-        description="Inference procedures: asymptotic / bootstrap / placebo.",
-    )
-    n_bootstrap: int = Field(
-        default=1000, ge=100,
-        description="Number of bootstrap replicates (Web Appendix F).",
-    )
-    bootstrap_seed: int = Field(
-        default=0,
-        description="Seed for the bootstrap RNG.",
-    )
-
-    @model_validator(mode="after")
-    def check_fma_params(cls, values: Any) -> Any:
-        allowed = {"asymptotic", "bootstrap", "placebo"}
-        for m in values.inference_methods:
-            if m not in allowed:
-                raise MlsynthConfigError(
-                    f"inference_methods entry {m!r} is not one of "
-                    f"{sorted(allowed)}."
-                )
-        return values
-
-class PDAConfig(BaseEstimatorConfig):
-    """Configuration for the Panel Data Approach (PDA) estimator."""
-    method: str = Field(default="fs", description="Type of PDA to use: 'LASSO', 'l2', or 'fs'.", pattern="^(LASSO|l2|fs)$")
-    methods: Optional[List[str]] = Field(default=None, description="Optional list of PDA variants to run; overrides `method` when set.")
-    tau: Optional[float] = Field(default=None, description="User-specified treatment effect value (used as tau_l2 for 'l2' method).")
-    alpha: float = Field(default=0.05, gt=0.0, lt=1.0, description="Significance level for confidence intervals and ATE inference.")
-    fs_intercept: bool = Field(default=False, description="Forward-selection only: include a constant in the donor regression. False (default) matches Shi & Huang's simulation (no intercept, valid size on mean-zero factor data); True matches the released fsPDA R package (intercept, for panels with genuine level differences).")
-
 # FDIDConfig has been relocated to mlsynth/utils/fdid_helpers/config.py
 # (co-located with the estimator's helpers). It is re-exported from this
 # module for backward compatibility via the module-level __getattr__ below.
@@ -443,672 +348,6 @@ class PDAConfig(BaseEstimatorConfig):
     # in the implementation. If 'save' functionality specific to GSC is re-added,
     # it might need to be defined here if different from BaseEstimatorConfig.save.
 
-
-
-class CLUSTERSCConfig(BaseEstimatorConfig):
-    """Configuration for the Cluster-based Synthetic Control (CLUSTERSC) estimator.
-
-    CLUSTERSC packages two robust synthetic-control families behind a
-    single interface:
-
-    * **PCR-RSC** -- SVD-based donor clustering plus Principal Component
-      Regression weight estimation (Amjad, Shah, Shen 2018; Agarwal,
-      Shah, Shen, Song 2021). Frequentist QP via :mod:`cvxpy` or
-      Bayesian Robust Synthetic Control posterior (Amjad, Shah, Shen 2018).
-    * **RPCA-SC** -- robust low-rank donor denoising (PCP -- Candes,
-      Li, Ma, Wright 2011; or HQF -- Wang, Li, So, Liu 2023) followed
-      by simplex SCM weights.
-
-    Parameters
-    ----------
-    method : {"pcr", "rpca", "both"}
-        Estimation family to run. ``"both"`` runs PCR and RPCA in
-        parallel; ``primary`` selects which one exposes ``att`` /
-        ``counterfactual`` / ``gap`` on the result object.
-    primary : {"pcr", "rpca"}
-        Which fit drives the result aliases when ``method = "both"``.
-    pcr_objective : {"OLS", "SIMPLEX"}
-        Inner SCM weight objective for the PCR family.
-    clustering : bool
-        Whether to apply SVD-based donor clustering before the PCR fit.
-    estimator : {"frequentist", "bayesian"}
-        Frequentist QP versus Bayesian posterior for the PCR family.
-    rpca_method : {"PCP", "HQF"}
-        Robust-PCA decomposition for the RPCA family.
-    lambda_penalty, p, q : float or None
-        Elastic-net-style regularization knobs forwarded to the
-        PCR inner solver.
-    alpha : float
-        Two-sided level for the Bayesian credible interval (PCR
-        estimator = "bayesian" only).
-    """
-
-    method: Literal["pcr", "rpca", "both"] = Field(
-        default="pcr",
-        description="Family to run: pcr, rpca, or both.",
-    )
-    primary: Literal["pcr", "rpca"] = Field(
-        default="pcr",
-        description="Which fit drives the result aliases when method='both'.",
-    )
-    pcr_objective: Literal["OLS", "SIMPLEX"] = Field(
-        default="OLS",
-        description="Inner SCM weight objective for the PCR family.",
-    )
-    clustering: bool = Field(
-        default=True,
-        description="Whether to apply SVD-based donor clustering before the PCR fit.",
-    )
-    estimator: Literal["frequentist", "bayesian"] = Field(
-        default="frequentist",
-        description="Frequentist QP or Bayesian posterior for the PCR family.",
-    )
-    rpca_method: Literal["PCP", "HQF"] = Field(
-        default="PCP",
-        description="Robust-PCA decomposition for the RPCA family.",
-    )
-    lambda_penalty: Optional[float] = Field(
-        default=None, ge=0.0,
-        description="Elastic-net regularization strength (lambda).",
-    )
-    p: Optional[float] = Field(
-        default=None, ge=0.0,
-        description="Norm parameter p for the elastic-net penalty.",
-    )
-    q: Optional[float] = Field(
-        default=None, ge=0.0,
-        description="Secondary norm parameter q (mixed-norm).",
-    )
-    alpha: float = Field(
-        default=0.05, gt=0.0, lt=1.0,
-        description="Two-sided credible-interval level (Bayesian PCR only).",
-    )
-    rank: Optional[int] = Field(
-        default=None, ge=1,
-        description="Explicit HSVT truncation rank `r` (Rho et al. 2025, Algorithm 2). "
-                    "If unset, rank_method controls selection.",
-    )
-    rank_method: Literal["cumvar", "fixed", "usvt"] = Field(
-        default="cumvar",
-        description="Rank-selection rule: 'cumvar' (paper default; smallest r "
-                    "with cumulative spectral energy >= cumvar_threshold), 'fixed' "
-                    "(use explicit `rank`), or 'usvt' (Chatterjee 2015 / Donoho-Gavish).",
-    )
-    cumvar_threshold: float = Field(
-        default=0.95, gt=0.0, le=1.0,
-        description="Cumulative-variance target when rank_method='cumvar'. "
-                    "Paper Section 6.1 uses 0.95.",
-    )
-    standardize_for_rank: bool = Field(
-        default=True,
-        description="If True (default), the cumvar / USVT rank rules operate on the "
-                    "column-standardised donor matrix so the spectral comparison "
-                    "is not dominated by uncentered level information. The HSVT "
-                    "step itself still consumes the raw matrix.",
-    )
-    project_denoised: bool = Field(
-        default=False,
-        description="If True, project f_hat through the HSVT-denoised donor matrix "
-                    "in both pre and post periods (Rho et al. 2025 Algorithm 4 "
-                    "Step 5, paper-strict). Default False uses the raw donor "
-                    "outcomes for the counterfactual, matching the practical "
-                    "behaviour of Amjad-Shah-Shen (2018) and legacy mlsynth.",
-    )
-    k_clusters: Optional[int] = Field(
-        default=None, ge=1,
-        description="Number of clusters for Algorithm 3. If unset, the silhouette "
-                    "coefficient picks k in [2, k_max].",
-    )
-    k_max: int = Field(
-        default=8, ge=2,
-        description="Upper bound for the silhouette-driven k-search.",
-    )
-    n_bayes_samples: int = Field(
-        default=1000, ge=1,
-        description="Posterior sample count for Bayesian PCR credible bands.",
-    )
-    random_state: int = Field(
-        default=0,
-        description="Seed forwarded to k-means and the Bayesian sampler.",
-    )
-    compute_shen_ci: bool = Field(
-        default=True,
-        description="When True (default) and the frequentist OLS PCR path is used, "
-                    "compute Shen-Ding-Sekhon-Yu (2023) per-period and ATT confidence "
-                    "intervals under three sources of randomness (HZ / VT / DR).",
-    )
-    shen_variance: Literal["homoskedastic", "jackknife", "hrk"] = Field(
-        default="homoskedastic",
-        description="Variance estimator for the Shen et al. (2023) CIs. "
-                    "'homoskedastic' (paper eq 19, default), 'jackknife', or 'hrk' "
-                    "(Hartley-Rao-Kish; only valid when max(1 - diag(H_perp)) < 1/2 "
-                    "for both projections).",
-    )
-    fpca_cumvar: float = Field(
-        default=0.95, gt=0.0, le=1.0,
-        description="Cumulative-variance target for the FPCA truncation in "
-                    "RPCA-SC Step 1 (Bayani 2021 recommends >= 0.95).",
-    )
-    pcp_lambda: Optional[float] = Field(
-        default=None, ge=0.0,
-        description="PCP sparsity penalty (Candes et al. 2011). Defaults to "
-                    "1/sqrt(max(N, T)) when unset.",
-    )
-    pcp_mu: Optional[float] = Field(
-        default=None, gt=0.0,
-        description="PCP augmented-Lagrangian penalty mu. Defaults to "
-                    "N*T / (4 * sum |Y|) (Bayani 2021).",
-    )
-    pcp_max_iter: int = Field(
-        default=1000, ge=1,
-        description="Maximum ADMM iterations for the PCP solver.",
-    )
-    pcp_tol: float = Field(
-        default=1e-9, gt=0.0,
-        description="Convergence tolerance for the PCP solver (relative to ||Y||_F).",
-    )
-    hqf_rank: Optional[int] = Field(
-        default=None, ge=1,
-        description="Explicit factorisation rank for HQF. If unset, picked by "
-                    "hqf_cumvar (Bayani 2021 uses 0.999).",
-    )
-    hqf_cumvar: float = Field(
-        default=0.999, gt=0.0, le=1.0,
-        description="Cumulative-variance target for HQF rank selection when "
-                    "hqf_rank is None.",
-    )
-    hqf_lambda: Optional[float] = Field(
-        default=None, ge=0.0,
-        description="Tikhonov factor for HQF (Wang et al. 2023). Defaults to "
-                    "1/sqrt(max(m, n)).",
-    )
-    hqf_ip: float = Field(
-        default=1.0, gt=0.0,
-        description="HQF noise-scale adaptation factor (Bayani 2021 default 1.0).",
-    )
-    hqf_max_iter: int = Field(
-        default=1000, ge=1,
-        description="Maximum iterations for the HQF solver.",
-    )
-    cv_lambda: bool = Field(
-        default=False,
-        description="Leave-one-time-period-out CV for PCP's sparsity penalty "
-                    "lambda. Sweeps cv_lambda_multipliers x Candes default, picks "
-                    "the value with lowest held-out NNLS MSE. On the California "
-                    "Prop 99 panel this halves pre-RMSE vs the Candes default.",
-    )
-    cv_hqf_rank: bool = Field(
-        default=False,
-        description="Leave-one-time-period-out CV for HQF's factorisation rank. "
-                    "Sweeps integer ranks in [1, min(J, T0-1)] and picks the rank "
-                    "with lowest held-out NNLS MSE.",
-    )
-    compute_cft_pi: bool = Field(
-        default=False,
-        description="Cattaneo-Feng-Titiunik (2021) prediction intervals for the "
-                    "RPCA-SC fit. Two-component (in-sample bootstrap + out-of-sample "
-                    "Hoeffding bound). Default False because it requires `cft_sims` "
-                    "full refits of the pipeline (~0.5s each).",
-    )
-    cft_sims: int = Field(
-        default=200, ge=10,
-        description="Number of bootstrap draws for the CFT in-sample component.",
-    )
-    cft_alpha: float = Field(
-        default=0.05, gt=0.0, lt=1.0,
-        description="Two-sided level for the CFT prediction intervals.",
-    )
-    cft_e_method: Literal["gaussian"] = Field(
-        default="gaussian",
-        description="Out-of-sample method for the CFT component. Currently "
-                    "only 'gaussian' (Hoeffding sub-Gaussian bound) is supported.",
-    )
-
-    @model_validator(mode="before")
-    @classmethod
-    def _normalise_legacy_fields(cls, data: Any) -> Any:
-        """Accept the legacy field names and uppercase method strings."""
-        if not isinstance(data, dict):
-            return data
-        # objective -> pcr_objective
-        if "objective" in data and "pcr_objective" not in data:
-            data["pcr_objective"] = data.pop("objective")
-        else:
-            data.pop("objective", None)
-        # cluster -> clustering
-        if "cluster" in data and "clustering" not in data:
-            data["clustering"] = data.pop("cluster")
-        else:
-            data.pop("cluster", None)
-        # Frequentist -> estimator
-        if "Frequentist" in data and "estimator" not in data:
-            data["estimator"] = (
-                "frequentist" if data.pop("Frequentist") else "bayesian"
-            )
-        else:
-            data.pop("Frequentist", None)
-        # ROB / Robust -> rpca_method
-        for k in ("ROB", "Robust"):
-            if k in data and "rpca_method" not in data:
-                data["rpca_method"] = data.pop(k)
-            else:
-                data.pop(k, None)
-        # Uppercase method tags
-        if "method" in data and isinstance(data["method"], str):
-            mapping = {"PCR": "pcr", "RPCA": "rpca", "BOTH": "both"}
-            data["method"] = mapping.get(data["method"], data["method"])
-        return data
-
-
-
-class DSCConfig(BaseEstimatorConfig):
-    """Configuration for the Distributional Synthetic Control (DSC) estimator.
-
-    DSC (Gunsilius 2023; asymptotic theory: Zhang, Zhang & Zhang 2026)
-    fits simplex-constrained weights on the *quantile functions* of
-    donor outcomes to reconstruct the treated unit's counterfactual
-    *outcome distribution*. Unlike the other mlsynth estimators, DSC
-    expects micro-level panel data: each ``(unit, time)`` cell carries
-    multiple individual observations, supplied as one row per
-    individual in the input DataFrame.
-
-    Parameters
-    ----------
-    M : int, optional
-        Number of quantile-grid points used to approximate the
-        2-Wasserstein loss. If ``None``, defaults to
-        ``max(200, min_cell_size)`` (Zhang et al. 2026 suggest
-        :math:`M = C n` for a constant ``C >= 1``).
-    grid_method : {"halton", "sobol", "uniform"}
-        Sampling rule for the quantile grid. ``"halton"`` (default)
-        and ``"sobol"`` are quasi-Monte Carlo with Koksma-Hlawka error
-        :math:`O(\\log M / M)`; ``"uniform"`` is i.i.d. with
-        :math:`O(M^{-1/2})` error.
-    lambda_method : {"uniform", "recency"}
-        Default rule for the pre-period aggregation weights
-        :math:`\\lambda_t`. Ignored when ``lambda_weights`` is set.
-    lambda_decay : float
-        Geometric decay factor for ``lambda_method="recency"``.
-        Default ``0.9``.
-    lambda_weights : sequence of float, optional
-        Caller-supplied length-``T0`` aggregation weights (must be
-        non-negative and sum to 1). Useful for Arkhangelsky et al.
-        (2021) SDiD-style time weights computed externally.
-    qte_quantiles : sequence of float, optional
-        Quantile grid in ``(0, 1)`` at which to report the QTE. If
-        ``None``, an evenly spaced grid of ``n_qte_points`` quantiles
-        is used.
-    n_qte_points : int
-        Length of the default QTE grid when ``qte_quantiles`` is None.
-        Default 99 (every percentile from 0.01 to 0.99).
-    random_state : int
-        Seed forwarded to the QMC quantile-grid sampler.
-    """
-
-    M: Optional[int] = Field(
-        default=None, ge=50,
-        description="Number of quantile-grid points used to approximate the "
-                    "2-Wasserstein loss. Defaults to max(200, min cell size).",
-    )
-    grid_method: Literal["halton", "sobol", "uniform"] = Field(
-        default="halton",
-        description="Quantile-grid sampling rule (QMC by default).",
-    )
-    lambda_method: Literal["uniform", "recency"] = Field(
-        default="uniform",
-        description="Default rule for pre-period aggregation weights lambda_t.",
-    )
-    lambda_decay: float = Field(
-        default=0.9, gt=0.0, le=1.0,
-        description="Geometric decay for lambda_method='recency'.",
-    )
-    lambda_weights: Optional[List[float]] = Field(
-        default=None,
-        description="Caller-supplied length-T0 aggregation weights "
-                    "(non-negative, sum to 1).",
-    )
-    qte_quantiles: Optional[List[float]] = Field(
-        default=None,
-        description="Explicit quantile grid in (0, 1) for QTE reporting.",
-    )
-    n_qte_points: int = Field(
-        default=99, ge=1,
-        description="Length of the default QTE grid when qte_quantiles is None.",
-    )
-    random_state: int = Field(
-        default=0,
-        description="Seed forwarded to the QMC sampler.",
-    )
-    compute_inference: bool = Field(
-        default=False,
-        description="Run the Gunsilius (2023) placebo permutation test "
-                    "(Algorithm 1): refit DSC with every donor treated as a "
-                    "placebo and rank the real unit's post-period Wasserstein "
-                    "distance. Costs J extra pre-period refits.",
-    )
-    inference_grid_points: int = Field(
-        default=200, ge=2,
-        description="Number of quantiles used to evaluate the squared "
-                    "2-Wasserstein distances in the placebo permutation test.",
-    )
-
-
-
-
-
-
-class MUSCConfig(BaseEstimatorConfig):
-    """Configuration for the Modified Unbiased Synthetic Control (MUSC) estimator.
-
-    MUSC is the Bottmer, Imbens, Spiess & Warnick (2024 JBES)
-    modification of the synthetic control method: a single column-
-    sums-to-zero linear restriction is added to the standard SC
-    quadratic programme so that the resulting estimator is exactly
-    unbiased under random assignment of which unit is treated (the
-    paper's Lemma 1).
-    """
-
-    alpha: float = Field(
-        default=0.05, gt=0.0, lt=1.0,
-        description=(
-            "Two-sided significance level for both the Normal-approximation "
-            "and the randomization-based CI surfaced on the MUSC inference "
-            "object. Defaults to 0.05 (95% CI)."
-        ),
-    )
-    run_inference: bool = Field(
-        default=True,
-        description=(
-            "Whether to compute the Proposition 1 unbiased variance "
-            "estimator (eq. 3.3) and the Section 3.5 randomization-based "
-            "confidence interval. The randomization CI requires one "
-            "leave-one-out QP refit per donor unit, so disabling it is "
-            "useful for exploratory fits on large panels."
-        ),
-    )
-    solver: Optional[str] = Field(
-        default=None,
-        description=(
-            "cvxpy solver name forwarded to the MUSC quadratic programme. "
-            "Defaults to None, which routes to cvxpy's default for the "
-            "problem class (CLARABEL on cvxpy >= 1.4)."
-        ),
-    )
-
-
-
-class MASCConfig(BaseEstimatorConfig):
-    """Configuration for the MASC estimator (Kellogg et al. 2021).
-
-    Kellogg, Mogstad, Pouliot & Torgovitsky (2021). *Combining
-    Matching and Synthetic Control to Trade Off Biases from
-    Extrapolation and Interpolation.* JASA 116(536), 1804-1816.
-    The estimator forms a convex combination ``phi * matching +
-    (1 - phi) * SC`` with the number of neighbours ``m`` and the
-    weight ``phi`` jointly chosen by rolling-origin cross-validation.
-    """
-
-    covariates: Optional[List[str]] = Field(
-        default=None,
-        description=(
-            "Optional predictor columns added to the SC objective in "
-            "place of (or alongside) pre-period outcomes. Each covariate "
-            "is row-standardised across all units before the QP -- the "
-            "FSCM/MAREX-style ``V = diag(1 / var)`` preconditioning that "
-            "matches Abadie's default heuristic V. Per-fold aggregation "
-            "is over the fold-specific pre-window unless "
-            "``covariate_windows`` overrides it."
-        ),
-    )
-    covariate_windows: Optional[dict] = Field(
-        default=None,
-        description=(
-            "Optional inclusive ``(start, end)`` aggregation window per "
-            "covariate label. Covariates without a window are averaged "
-            "over each fold's full pre-period."
-        ),
-    )
-    m_grid: Optional[List[int]] = Field(
-        default=None,
-        description=(
-            "Candidate nearest-neighbour counts for cross-validation. "
-            "Defaults to ``1..J`` (all values), mirroring the R "
-            "reference's default."
-        ),
-    )
-    min_preperiods: Optional[int] = Field(
-        default=None, ge=2,
-        description=(
-            "Smallest fold length used in cross-validation. Folds run "
-            "from ``min_preperiods`` to ``treatment_period - 2``. "
-            "Defaults to ``ceil(treatment_period / 2)`` per the R "
-            "reference. Mutually exclusive with ``set_f``."
-        ),
-    )
-    set_f: Optional[List[int]] = Field(
-        default=None,
-        description=(
-            "Explicit list of fold lengths to use (each integer is the "
-            "last pre-treatment period included in that fold's "
-            "training window). Mutually exclusive with "
-            "``min_preperiods``."
-        ),
-    )
-    fold_weights: Optional[List[float]] = Field(
-        default=None,
-        description=(
-            "Optional positive relative weights for each fold (length "
-            "matches the fold count). Normalised to sum to 1. Defaults "
-            "to equal weights."
-        ),
-    )
-    forecast_minlength: int = Field(
-        default=1, ge=1,
-        description=(
-            "First post-fold period used for forecast error "
-            "evaluation. Fold ``f`` forecasts period ``f + "
-            "forecast_minlength``."
-        ),
-    )
-    forecast_maxlength: int = Field(
-        default=1, ge=1,
-        description=(
-            "Last post-fold period used for forecast error "
-            "evaluation. Forecast window is capped at "
-            "``treatment_period - 1``."
-        ),
-    )
-    solver: Optional[str] = Field(
-        default=None,
-        description=(
-            "cvxpy solver name forwarded to the SC simplex QP. "
-            "Defaults to CLARABEL when unset."
-        ),
-    )
-
-
-
-class SpSyDiDConfig(BaseEstimatorConfig):
-    """Configuration for the Spatial Synthetic Difference-in-Differences estimator.
-
-    Serenini & Masek (2024). *"Spatial Synthetic
-    Difference-in-Differences,"* SSRN 4736857. Extends SDID
-    (Arkhangelsky et al. 2021) with a spatial spillover term so the
-    estimator separates the direct ATT from the indirect (spillover)
-    effect on units exposed via the spatial weight matrix :math:`W`.
-
-    Parameters
-    ----------
-    spatial_matrix : np.ndarray
-        Square :math:`N \\times N` spatial weight matrix. Rows /
-        columns must align with ``unit_order`` (or
-        ``sorted(df[unitid].unique())`` if ``unit_order`` is None).
-        Use the helpers in
-        :mod:`mlsynth.utils.spsydid_helpers.spatial` to build ``W``
-        from coordinates (k-NN, inverse distance) or from an adjacency
-        list (queen / rook contiguity).
-    unit_order : list, optional
-        Canonical ordering of unit ids matching the rows / columns of
-        ``spatial_matrix``. If ``None`` (default), units are ordered
-        by ``sorted(df[unitid].unique())``.
-    row_standardize_spatial : bool
-        Row-standardise ``W`` internally before computing exposure.
-        Default True. Skip when the caller has already standardised.
-    """
-
-    spatial_matrix: Any = Field(
-        ...,
-        description="Square (N, N) spatial weight matrix as a numpy array.",
-    )
-    unit_order: Optional[List[Any]] = Field(
-        default=None,
-        description="Canonical ordering of unit ids matching the rows / columns "
-                    "of spatial_matrix. Defaults to sorted(unique unit ids).",
-    )
-    row_standardize_spatial: bool = Field(
-        default=True,
-        description="Row-standardise the spatial matrix internally so each row "
-                    "of W sums to 1.",
-    )
-
-
-class MCNNMConfig(BaseEstimatorConfig):
-    """Configuration for the MC-NNM estimator.
-
-    Athey, Bayati, Doudchenko, Imbens & Khosravi (2021), *"Matrix
-    Completion Methods for Causal Panel Data Models"* (JASA). Imputes the
-    treated cells of the outcome matrix via nuclear-norm-regularised
-    low-rank matrix completion with unregularised two-way fixed effects
-    (SOFT-IMPUTE, threshold chosen by cross-validation). Inherits the
-    standard ``df`` / ``outcome`` / ``treat`` / ``unitid`` / ``time``
-    interface.
-
-    Parameters
-    ----------
-    estimate_unit_fe : bool
-        Estimate (unregularised) unit fixed effects. Default True.
-    estimate_time_fe : bool
-        Estimate (unregularised) time fixed effects. Default True.
-    n_lambda : int
-        Number of candidate singular-value thresholds in the CV grid.
-    n_folds : int
-        Cross-validation folds over the observed cells.
-    inference : bool
-        Run a leave-one-control jackknife for the ATT SE / CI. Default
-        False (it refits the model once per control unit).
-    alpha : float
-        Two-sided level for the jackknife confidence interval.
-    random_state : int
-        Seed for the CV fold assignment.
-    """
-
-    estimate_unit_fe: bool = Field(
-        default=True, description="Estimate unregularised unit fixed effects.")
-    estimate_time_fe: bool = Field(
-        default=True, description="Estimate unregularised time fixed effects.")
-    n_lambda: int = Field(
-        default=40, ge=2,
-        description="Number of candidate thresholds in the CV grid.")
-    n_folds: int = Field(
-        default=5, ge=2, description="Cross-validation folds over observed cells.")
-    inference: bool = Field(
-        default=False,
-        description="Run a leave-one-control jackknife for the ATT SE/CI.")
-    alpha: float = Field(
-        default=0.05, gt=0.0, lt=1.0,
-        description="Two-sided level for the jackknife confidence interval.")
-    random_state: int = Field(
-        default=0, description="Seed for the CV fold assignment.")
-
-
-class MSQRTConfig(BaseEstimatorConfig):
-    """Configuration for the MSQRT estimator.
-
-    Shen, Song & Abadie (2025), *"Efficiently Learning Synthetic Control
-    Models for High-dimensional Disaggregated Data"* (arXiv:2510.22828).
-    Stacks all treated units into one matrix regression ``Y = X Theta + E``
-    and fits the donor-weight matrix by Multivariate Square-root Lasso
-    (nuclear-norm loss + element-wise L1), with the L1 penalty chosen by
-    rolling-origin cross-validation on the pre-period. Assumes a block design
-    (all treated units adopt at the same period) and inherits the standard
-    ``df`` / ``outcome`` / ``treat`` / ``unitid`` / ``time`` interface.
-
-    Parameters
-    ----------
-    lambda_ : float, optional
-        Fixed L1 penalty. If None (default), selected by cross-validation.
-    n_lambda : int
-        Number of log-spaced candidate penalties in the CV grid.
-    cv_initial_train, cv_val_window, cv_step, cv_folds : int, optional
-        Rolling-origin CV schedule overrides. Adaptive defaults
-        (scaled to the pre-period length) are used when left None.
-    inference : bool
-        Attach CFPT/scpi prediction intervals (Cattaneo, Feng, Palomba &
-        Titiunik 2025) for all four predictands plus simultaneous bands. For
-        MSQRT only the out-of-sample error is modelled. Default False.
-    alpha : float
-        Total miscoverage level for the intervals (default 0.1 -> 90%).
-    time_dependence : {"iid", "general"}
-        Time-averaging assumption for the time-averaged predictands (TAUS,
-        TAUA). ``"iid"`` (default) shrinks the band by ``sqrt(L)``;
-        ``"general"`` makes no serial-dependence assumption.
-    """
-
-    lambda_: Optional[float] = Field(
-        default=None, gt=0.0,
-        description="Fixed L1 penalty; None selects it by cross-validation.")
-    n_lambda: int = Field(
-        default=15, ge=2,
-        description="Number of log-spaced candidate penalties in the CV grid.")
-    cv_initial_train: Optional[int] = Field(
-        default=None, ge=2,
-        description="Initial training window for rolling-origin CV.")
-    cv_val_window: Optional[int] = Field(
-        default=None, ge=1,
-        description="Validation window length for rolling-origin CV.")
-    cv_step: Optional[int] = Field(
-        default=None, ge=1, description="Step between successive CV folds.")
-    cv_folds: Optional[int] = Field(
-        default=None, ge=1, description="Maximum number of CV folds.")
-    inference: bool = Field(
-        default=False,
-        description="Attach CFPT/scpi prediction intervals to the predictands.")
-    alpha: float = Field(
-        default=0.1, gt=0.0, lt=1.0,
-        description="Total miscoverage level for the prediction intervals.")
-    time_dependence: Literal["iid", "general"] = Field(
-        default="iid",
-        description="Time-averaging assumption for time-averaged predictands.")
-
-
-class SSCConfig(BaseEstimatorConfig):
-    """Configuration for the SSC (Staggered Synthetic Control) estimator.
-
-    Cao, Lu & Wu (2026), *"Synthetic Control Inference for Staggered
-    Adoption"* (The Econometrics Journal). Models each unit's untreated
-    outcome as an intercept plus a simplex synthetic control on all other
-    units (not-yet-treated units are valid donors), jointly estimates every
-    unit x time effect by GLS, and reports event-time / overall ATT with
-    Andrews (2003) end-of-sample stability inference. Targets staggered
-    adoption with a long pre-period (large ``T``, moderate ``N``, small
-    ``S``). Inherits the standard ``df`` / ``outcome`` / ``treat`` /
-    ``unitid`` / ``time`` interface.
-
-    Parameters
-    ----------
-    inference : bool
-        Attach Andrews end-of-sample bands and p-values to the event-time and
-        overall ATT. Default True.
-    alpha : float
-        Two-sided level for the bands (default 0.1 -> 90% band).
-    """
-
-    inference: bool = Field(
-        default=True,
-        description="Attach Andrews end-of-sample bands/p-values to the ATTs.")
-    alpha: float = Field(
-        default=0.1, gt=0.0, lt=1.0,
-        description="Two-sided level for the end-of-sample bands.")
 
 
 class RMSIConfig(BaseEstimatorConfig):
@@ -4203,6 +3442,16 @@ class DesignResult(MlsynthResult):
 # BaseEstimatorConfig from here).
 # ---------------------------------------------------------------------------
 _RELOCATED_CONFIGS = {
+    "FMAConfig": "mlsynth.utils.fma_helpers.config",
+    "PDAConfig": "mlsynth.utils.pda_helpers.config",
+    "CLUSTERSCConfig": "mlsynth.utils.clustersc_helpers.config",
+    "DSCConfig": "mlsynth.utils.dsc_helpers.config",
+    "MUSCConfig": "mlsynth.utils.musc_helpers.config",
+    "MASCConfig": "mlsynth.utils.masc_helpers.config",
+    "SpSyDiDConfig": "mlsynth.utils.spsydid_helpers.config",
+    "MCNNMConfig": "mlsynth.utils.mcnnm_helpers.config",
+    "MSQRTConfig": "mlsynth.utils.msqrt_helpers.config",
+    "SSCConfig": "mlsynth.utils.ssc_helpers.config",
     "VanillaSCConfig": "mlsynth.utils.vanillasc_helpers.config",
     "FDIDConfig": "mlsynth.utils.fdid_helpers.config",
     "TSSCConfig": "mlsynth.utils.tssc_helpers.config",
